Simon: "Wo bin ich?"

+++ src/src/Main.java
@@ -12,11 +12,11 @@ public class Main {
 
         do {
             System.out.println("Geben Sie eine Zahl ein");
-            while (!benutzerEingabe.hasNextInt()) {
+            while (!benutzerEingabe.hasNextDouble()) {
                 System.out.println("Das ist keine Zahl");
                 benutzerEingabe.next();
             }
-            Zahl = benutzerEingabe.nextInt();
+            Zahl = benutzerEingabe.nextDouble();
         } while (Zahl <= 0);
 
         System.out.println("Danke, wir arbeiten also mit dieser Zahl: " + Zahl);
@@ -28,25 +28,25 @@ public class Main {
         switch(rechenart){
             case "addieren":
                 System.out.println("Welche Zahl soll zu " +Zahl +" addiert werden?");
-                double ZahlAdd = benutzerEingabe.nextInt();
+                double ZahlAdd = benutzerEingabe.nextDouble();
                 System.out.println("Das Ergebnis ist: " + rn.addition(Zahl, ZahlAdd));
                 break;
 
             case "subtrahieren":
                 System.out.println("Welche Zahl soll von " +Zahl +" subtrahiert werden?");
-                double ZahlSub = benutzerEingabe.nextInt();
+                double ZahlSub = benutzerEingabe.nextDouble();
                 System.out.println("Das Ergebnis ist: " + rn.dekrement(Zahl, ZahlSub));
                 break;
 
             case "multiplizieren":
                 System.out.println("Mit welcher Zahl soll " + Zahl + " multipliziert werden?");
-                double ZahlMulti = benutzerEingabe.nextInt();
+                double ZahlMulti = benutzerEingabe.nextDouble();
                 System.out.println("Das Ergebnis ist: " + rn.faktor(Zahl, ZahlMulti));
                 break;
 
             case "dividieren":
                 System.out.println("Durch welche Zahl soll " + Zahl + " dividiert werden?");
-                double ZahlDivi = benutzerEingabe.nextInt();
+                double ZahlDivi = benutzerEingabe.nextDouble();
                 System.out.println("Das Ergebnis ist: " + rn.division(Zahl, ZahlDivi));
                 break;
